requirements.txt
aiofiles==24.1.0
aiogram==3.19.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
asyncpg==0.30.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
django-jazzmin==3.0.1
djangorestframework==3.16.0
frozenlist==1.5.0
idna==3.10
magic-filter==1.0.12
multidict==6.2.0
pillow==11.1.0
propcache==0.3.1
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.13.0
urllib3==2.3.0
yarl==1.18.3
